replace plugin API with build/serialization steps
